feat: Complete Map Center & Cloud Default feature with backend integration - TIEMPO-312
This commit completes Phases 1 & 2 of the map center location feature:

Phase 1 - UI Improvements:
- Refactored MapCenterModal UI to clearly distinguish Session vs Cloud Default
- Auth-aware button layout: logged-in users see "Save as Default", anonymous see "Sign Up"
- Improved mobile responsiveness with proper loading indicators
- Added helpful alert messages explaining temporary vs permanent locations

Phase 2 - Backend Integration:
- Integrated with Firebase auth for token management
- Connected to Azure Functions API endpoints for location save/load
- Implemented proper error handling with fresh token refresh
- Location preferences now persist across devices via MongoDB
- Cleaned up UnifiedLocationModal (no longer needed)

Architecture:
- GeoLocationContext manages all location state and backend calls
- MapCenterModal handles UI and delegates to context
- UserSettingsLocationPreferences manages permanent saved locations
- All location data flows through unified context API

Tested successfully with real user authentication and data persistence.
